+++ library/src/main/java/com/squareup/otto/Subscribe.java
@@ -22,7 +22,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Marks a method as an event handler, as used by {@link AnnotatedHandlerFinder} and {@link Bus}.
+ * Marks a method as an event subscriber, as used by {@link ReflectionFinder} and {@link Bus}.
  *
  * <p>The method's first (and only) parameter defines the event type.
  * <p>If this annotation is applied to methods with zero parameters or more than one parameter, the object containing