+++ library/src/main/java/com/squareup/otto/internal/Target.java
@@ -13,23 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package com.squareup.otto;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+package com.squareup.otto.internal;
 
 /**
- * Marks a method as an instance producer, as used by {@link ReflectionFinder} and {@link Bus}.
- * <p>
- * Otto infers the instance type from the annotated method's return type. Producer methods may return null when there is
- * no appropriate value to share. The calling {@link Bus} ignores such returns and posts nothing.
+ * Base class for generated {@link com.squareup.otto.Producer} and {@link com.squareup.otto.Subscriber} implementations
+ * which hold a reference to a target object.
  *
  * @author Jake Wharton
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Produce {
+public abstract class Target<T> {
+  protected final T target;
+
+  protected Target(T target) {
+    this.target = target;
+  }
 }