+++ library/src/test/java/com/squareup/otto/ThreadEnforcerTest.java
@@ -16,6 +16,7 @@
 
 package com.squareup.otto;
 
+import org.junit.Before;
 import org.junit.Test;
 
 import static junit.framework.Assert.assertFalse;
@@ -31,31 +32,29 @@ public class ThreadEnforcerTest {
     }
   }
 
-  @Test public void enforerCalledForRegister() {
-    RecordingThreadEnforcer enforcer = new RecordingThreadEnforcer();
-    Bus bus = new Bus(enforcer);
+  private Bus bus;
+  private RecordingThreadEnforcer enforcer;
 
+  @Before public void setUp() {
+    enforcer = new RecordingThreadEnforcer();
+    bus = new Bus(enforcer);
+  }
+
+  @Test public void enforerCalledForRegister() {
     assertFalse(enforcer.called);
     bus.register(this);
     assertTrue(enforcer.called);
   }
 
   @Test public void enforcerCalledForPost() {
-    RecordingThreadEnforcer enforcer = new RecordingThreadEnforcer();
-    Bus bus = new Bus(enforcer);
-
     assertFalse(enforcer.called);
     bus.post(this);
     assertTrue(enforcer.called);
   }
 
   @Test public void enforcerCalledForUnregister() {
-    RecordingThreadEnforcer enforcer = new RecordingThreadEnforcer();
-    Bus bus = new Bus(enforcer);
-
     assertFalse(enforcer.called);
     bus.unregister(this);
     assertTrue(enforcer.called);
   }
-
 }