+++ sample/src/main/java/com/squareup/otto/sample/LocationMapFragment.java
@@ -79,7 +79,7 @@ public class LocationMapFragment extends Fragment {
     }
   }
 
-  private static class ImageAvailableEvent {
+  public static class ImageAvailableEvent {
     public final Drawable image;
 
     ImageAvailableEvent(Drawable image) {